基础类的实现

+++ core/node.go
@@ -2,9 +2,9 @@ package core
 
 type Node struct {
 	Key          string
-	MaxThreshold int64
+	MaxThreshold int
 }
 
-func NewNode(key string, maxThreshold int64) *Node {
+func NewNode(key string, maxThreshold int) *Node {
 	return &Node{Key: key, MaxThreshold: maxThreshold}
 }
